LPS-198452 Move class DeepPaginationConfigurationWrapper to class ElasticsearchIndexSearcher
